fix: Adding Grid refactor and tauri/browser detection
